+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 ## 28-3-2024
 #### - **_CREATE_** `local_db.py` to handle all local database operations
 #### - **_CREATE_** `requirements.txt` file to list all python dependencies
-#### - **_PUSH_** to github with tag local-db
+#### - **_PUSH_** to GitHub with tag local-db
+#### - **_REPLACE_** mainly some of code inside `init_setup.py`
+* **Creating variable to check using cloud database or not**
+* **Did some adjustment on init setup saving data from firebase to local**
+#### - **_PUSH_** to GitHub with tag init-setup adjustment
 
 ## 27-3-2024
 #### - _**DONE**_ new-transactions form page
-#### - _**PUSH**_ to github with tag new-transactions-form
+#### - _**PUSH**_ to GitHub with tag new-transactions-form